menu pengaturan

+++ app/Http/Controllers/PengaturanController.php
@@ -7,6 +7,8 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 use Illuminate\Support\Facades\File;
 use Intervention\Image\Facades\Image as ResizeImage;
+use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\Facades\DB;
 
 class PengaturanController extends Controller
 {
@@ -24,32 +26,61 @@ class PengaturanController extends Controller
 
     public function update(Request $request, $id)
     {
-        $pengaturan = Pengaturan::find($id);
+        // Validator
+        $validator = Validator::make(
+            $request->all(),
+            [
+                'title' => 'required',
+                'content' => 'required',
+                'img' => 'image|mimes:jpeg,png,jpg|max:2048',
+            ],
+            [
+                'title.required' => 'Judul wajib diisi.',
+                'content.required' => 'Kontent wajib diisi.',
+                'img.image' => 'Img berupa gambar.',
+                'img.mimes' => 'Img harus berupa jpeg,png,jpg.',
+                'img.max' => 'Img Maksimal 2mb',
+            ],
+        );
 
-        if ($request['img']) {
-            $path = public_path('uploads/img/');
-            !is_dir($path) &&
-                mkdir($path, 0777, true);
-
-            // Process delete old img
-            $oldImg = $pengaturan->img;
-            File::delete($path . $oldImg);
-
-            // Process Uploads
-            $name = Str::slug($pengaturan->title, '-') . '.' . $request->img->extension();
-            ResizeImage::make($request->file('img'))
-                ->resize(1920, 1080)
-                ->save($path . $name);
+        // If validator fails.
+        if ($validator->fails()) {
+            return redirect()->back()->withInput($request->all())->withErrors($validator);
         }
 
-        $newUpdate = [
-            'title' => $request->title,
-            'text' => $request->text,
-            'img' => $name ?? null,
-        ];
 
-        $pengaturan->update($newUpdate);
+        // If validator success
+        DB::beginTransaction();
+        try {
+            $pengaturan = Pengaturan::find($id);
+            if ($request['img']) {
+                $path = public_path('uploads/img/');
+                !is_dir($path) &&
+                    mkdir($path, 0777, true);
 
-        return 'success';
+                // Process delete old img
+                $oldImg = $pengaturan->img;
+                File::delete($path . $oldImg);
+
+                // Process Uploads
+                $name = Str::slug($pengaturan->title, '-') . '.' . $request->img->extension();
+                ResizeImage::make($request->file('img'))
+                    ->resize(1920, 1080)
+                    ->save($path . $name);
+            }
+
+            $pengaturan->update([
+                'title' => $request->title,
+                'text' => $request->content,
+                'img' => $name ?? $pengaturan->img,
+            ]);
+
+            return redirect()->route('pengaturan.index')->with('success', 'Pengaturan berhasil diupdate.');
+        } catch (\Throwable $th) {
+            DB::rollBack();
+            return redirect()->route('pengaturan.index')->with('fails', 'Pengaturan gagal diupdate.');
+        } finally {
+            DB::commit();
+        }
     }
 }
